Refactor of OpenMPISettings and improved support for mpiexec (#242)
mpiexec, mpirun, and orterun are all synonyms in the modern MPI
standards. The mpirunSettings and mpirunStep have been refactored
to the more generic mpiSettings and mpiStep. Within these modules
the classes associated with mpiexec, mpirun, and orterun all
inherit from mpiexec (which historically was the first executable
method that was defined in the MPI standard).

One exception to this is that the slurm workload manager
provides a wrapper to mpiexec which wraps a subset of the
supported options to mpiexec so that they can be passed to Slurm.
In general, this is a very limited set of options and so we choose
not to support it. Instead, we check to see if mpiexec is actually
mpiexec.slurm and throw an UnsupportedError.

Another exception (which is now supported) is the Parallel
Application Launch Service used on more recent HPE/Cray HPC
platforms which also wraps mpiexec. Users of these platforms
should now use the new run settings class palsSettings and PALS
launchers for their workflows.

[ Commited by @ashao ]
[ Reviewed by @MattToast ]

Co-authored-by: Riccardo Balin <riccardo.balin@colorado.edu>